question-generator/pyproject.toml
[project]
name = "question-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "strands-agents>=1.10.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "mcp>=1.0.0",
]
